Fix: 비밀번호 변경 파일 폴더 이동 및 검증 코드 수정
