+++ services/user.js
@@ -85,41 +85,12 @@ export const logIn = async (email, password) => {
     throw error;
   }
 
-
   const [{ id }] = await userRepository.getUserIdByEmail(email);
   const token = jwt.sign({ id }, process.env.SECRET_KEY);
   return token;
 };
 
-const updatePassword = async (
-  userId,
-  password,
-  newPassword,
-  confirmNewPassword
-) => {
-  const userPassword = await userRepository.getUserPasswordbyId(userId);
-  const check = await bcrypt.compare(password, userPassword[0].password);
-  const salt = await bcrypt.genSalt();
-
-  if (!check) {
-    const error = new Error('비밀번호가 틀렸습니다.');
-    error.statusCode = 400;
-    throw error;
-  }
-  if (!(newPassword && confirmNewPassword)) {
-    const error = new Error('비밀번호가 일치하지 않습니다.');
-    error.statusCode = 400;
-    throw error;
-  }
-
-  const encryptPw = bcrypt.hash(newPassword, salt);
-
-  return await userRepository.updatePassword(userId, encryptPw);
- 
-};
-
 export const me = async userId => {
-  console.log(userId);
   const user = await userRepository.getUserIdbyId(userId);
   if (user.length === 0) {
     const error = new Error('User Not Found');
